Provide old name as initial input to interactive rename of perspective
In case the rename is intended to result in only a slightly modified
version of the old name, it is convenient to edit it rather than type
it out all the way from scratch.

In case of a rename to a completely different name, it should be easy
enough to first kill the initial input with something like `C-a k` or
`M-backspace` before typing it out.