add support for handling cloned indirect buffers
Currently, if a user clones a buffer (with e.g., 'clone-indirect-buffer'), the
new clone will not be added to the current workspace.

Add new customizable option to allow user to specify what should be done when an
indirect buffer is created via cloning. By default (t or non-nil), if a buffer
is created via cloning, it is added to the current perspective only if its
associated base-buffer is also in the current perspective. Option can be set to
nil to never automatically add clone indirect buffers, or to 'always to add a
clone to the current perspective regardless of whether the base-buffer is in the
current perspective.